Move string stuff to separate module
